LayoutTests/fast/js/delete-syntax-expected.txt

This test checks whether various forms of delete expression are allowed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS delete x is true
PASS delete window.x is true
PASS delete window["x"] is true
PASS delete (x) is true
PASS delete (window.x) is true
PASS delete (window["x"]) is true
PASS (y, delete x) is true
PASS delete ((x)) is true
PASS delete ((window.x)) is true
PASS delete ((window["x"])) is true
PASS delete (y, x) is true
PASS delete (true ? x : y) is true
PASS delete nonexistent is true
PASS delete window.nonexistent is true
PASS delete window["nonexistent"] is true
PASS delete (nonexistent) is true
PASS delete (window.nonexistent) is true
PASS delete (window["nonexistent"]) is true
PASS delete "x" is true
PASS delete (2 + 3) is true
PASS successfullyParsed is true

TEST COMPLETE
